requirements.txt
fastapi
uvicorn
langchain
PyPDF2
aiofiles
transformers
sentence-transformers
faiss-cpu
python-dotenv
